fix(core): widen Rect.area to Long and document the geometry types
Rect.area computed `width * height` as Int, which silently overflows to a
negative number past a ~46340-pixel side — reachable on a full-frame ROI
of a large image. Compute it as Long. Also add the Scaladoc that Point,
Size, Rect and Scalar were missing (they appear in nearly every
signature), including the BGR channel-order note on Scalar.

+++ core/src/scalacv/Geometry.scala
@@ -10,6 +10,9 @@ import org.opencv.core as cv
   * here — four ints — and it makes the results ordinary immutable Scala data that is safe to keep after the
   * Mat it came from is released.
   */
+/** A 2D point in pixel coordinates, origin top-left, `x` right and `y` down — the sub-pixel form OpenCV uses
+  * for feature and contour work, so both fields are `Double`.
+  */
 final case class Point(x: Double, y: Double):
   private[scalacv] def toCv: cv.Point = cv.Point(x, y)
 
@@ -19,18 +22,34 @@ final case class Point(x: Double, y: Double):
 final case class Point3(x: Double, y: Double, z: Double):
   private[scalacv] def toCv: cv.Point3 = cv.Point3(x, y, z)
 
+/** A width/height extent in pixels. Neither side may be negative — a zero extent is allowed (an empty size),
+  * a negative one throws.
+  */
 final case class Size(width: Double, height: Double):
   require(width >= 0 && height >= 0, s"a Size cannot be negative: ${width}x$height")
   private[scalacv] def toCv: cv.Size = cv.Size(width, height)
 
+/** An axis-aligned integer rectangle: top-left corner `(x, y)` and non-negative `width`/`height`. The origin
+  * may be negative (a region of interest can extend past the top-left of the image); the extent may not.
+  */
 final case class Rect(x: Int, y: Int, width: Int, height: Int):
   require(width >= 0 && height >= 0, s"a Rect cannot have negative extent: ${width}x$height")
-  def area: Int = width * height
+
+  /** The enclosed area in pixels. `Long`, because `width * height` overflows a signed `Int` past roughly a
+    * 46340-pixel side — an easy limit to hit on a full-frame ROI of a large image.
+    */
+  def area: Long = width.toLong * height
+
+  /** The top-left corner as a [[Point]]. */
   def topLeft: Point = Point(x.toDouble, y.toDouble)
+
+  /** The bottom-right corner as a [[Point]] — `(x + width, y + height)`, one past the last enclosed pixel. */
   def bottomRight: Point = Point((x + width).toDouble, (y + height).toDouble)
   private[scalacv] def toCv: cv.Rect = cv.Rect(x, y, width, height)
 
-/** A pixel value, in whatever channel order the Mat uses — OpenCV's default is BGR, not RGB. */
+/** A pixel value: up to four channel components, in whatever channel order the Mat uses. OpenCV's default is
+  * **BGR, not RGB**, so [[Scalar.Red]] is `Scalar(0, 0, 255)`. Unset channels default to `0`.
+  */
 final case class Scalar(v0: Double, v1: Double = 0, v2: Double = 0, v3: Double = 0):
   private[scalacv] def toCv: cv.Scalar = cv.Scalar(v0, v1, v2, v3)
